Working on drag-and-drop
Created HitBoxSprite class as wrapper for react-game-kit Sprite class.
Still had to modify Sprite class in order to create a React ref to the
internal native components because PanResponder is more computationally
efficient when using Direct Manipulation.

For now there is a bug where the locationX of the native event is
changing because internally, the react-game-kit Sprite classe changes
the x coordinate of the Image and then crops it in order to create the
effect of animation.

I will look into how circumvent this or at least normalize the
coordinate based on the current animation frame.

Also, added some reducer and action files for Redux because I realized
that the pet’s position should really be in Redux state because this
makes things much easier later when I want to implement other
HitBoxSprites to interact with (such as candy to eat, treadmills to
walk on, or just really any type of drag-and-drop “zone”). But that is
a problem for another day.