Added HitBoxSprite
Added wrapper class HitBoxSprite over modified Sprite class from
react-game-kit.

Has hooks for touchStart,touchMove, and touchEnd as well as hitBox
coordinate specification.

I will use this as a reusable component for other sprites such as food
and accessories.

I chose to create this class rather than using the MatterJS physics
engine that is recommended with react-game-kit because this combines
both hitboxes and drag-and-drop functionality. Also, an entire 2D
physics engine is way overkill for the drag-and-drop functionality that
I need.

That being said, I will keep MatterJS in mind in case I need more
physics functionality later in development.