Replaced react-native-router-flux with Navigator
I decided to go with React Native’s own Navigator because
react-native-router-flux was rendering slowly and it had some odd
behavior where it would render the Backpack scene then quickly revert
to the Store scene when the Store button was pressed.

After several hours of trying all kinds of variations of the API, I
just went with Navigator because my navigation needs are nothing short
of primitive. 3 scenes. Home at the root and 2 leaf node scenes, Store
and Inventory.

Still, Navigator has this *bug* where the user can click the navigation
button while the scene is being animated away and this causes the
onPress handler to re-push the same scene to the stack, thereby causing
a “Warning: flattenChildren(...): Encountered two children with the
same key.”

Will look into a possible workaround but for now it’s not a major
issue. I want to focus more on feature development for now to get beta
testing up and running ASAP.