Added Wellbeing Stats
Added a lower set of icons for the pet’s wellbeing statistics (health,
hunger, happiness, and wealth I guess).

Animations are working nicely. That was by far the hardest part of
this. I tried everything from modals to nested navigators but in the
end simply some component hierarchy and some Animated.Views were all it
took.

Next step is to actually have the values change over time, affect the
animation of the main pet Sprite and implement the Store and Backpack
screens as well as the PawPoints modal.

Overall, I am satisfied with the progress being made. The UI hiccups
here and there are not to my delight. But things should flow smoother
as I get used to which React Native components are better at solving
certain problems.