LoginForm not working
There is something really weird going on with TextInput in this version
of React/React Native, so I am just going to downgrade to the version
that I used when doing a Udemy React Native course because it worked
for me there. This is the latest commit before I attempt to do some
downgrading in case I need a full reset.