Fixed Router bug
Well, I was chasing this UI bug all day, but what is turned out to be
was that React Native’s Navigator was rendering a navigation bar even
at the Auth screen when all the navigation bar buttons and title are
null, so there was basically an invisible navigation bar blocking touch
events from propagating to the TextInput components.

I fixed it by simply adding a marginTop to the auth TextInput
components.

I also refactored some styling. Next, time to work on the Store and
Backpack screen and adding Redux state.