Added WellbeingStat Redux state and styled Store
I was able to write a short algorithm for adding Redux state to the
app. With it, I added the wellbeing statistics of the pet (happiness,
hunger, health, pawPoints) to the Redux state and it is working
beautifully. For now, some simple Timers are varying the statistics
over time.

Also, I added much better styling to the Store. Full-on game
functionality is not far away. :)