Basic "Real Time" Step Counter
So after digging through and out the Apple HealthKit documentation, I
finally found a single StackOverflow comment which recommended using
HKStatisticsCollectionQuery which, through some trial and error, I
managed to work into updating the step counter live within the React
Native app as soon as data comes in to HealthKit. I could use the Core
Motion CMPedometer module for even faster updates but really, the
implementation I have suites my needs just fine.

Now, I just have to tie the step count into in-game currency and think
of a system to track which steps have been counted and which steps are
left to exchange for in-game currency.

Then, I will have some more brainstorming sessions about the actual
game mechanics and how to make the app more engaging. Getting PawPoints
for walking is cool, but without an engaging in-game experience, it
won’t be fun. Back to the drawing board, but again, I am happy with the
progress.