Refactored State and added Shop Dialog
I removed some irrelevant action creators since they just returned the
default state anyway.

I also successfully completed a dialog box for the shop that opens when
a user chooses to inspect a piece of merchandise. It displays the
effects of the given item on the pet’s health, hunger, and mood along
with price, current PawPoints, and quantity already owned in the
Backpack.

It also comes with really nice colors and buttons.

Now I will work on the “placing” mechanic where a user chooses to
“place” an item from the backpack to the Pet and  then drag-and-drop
the Pet onto the item to activate its effects  on the Pet’s wellbeing.