Finished item placement mechanism
I finished the placement mechanism which allows a user to “place” an
item from the backpack into the Home screen and drag-n-drop it onto the
Pet (or vice versa) to “consume” the item and apply whatever effects
the item has. I am glad I made this mechanism pretty modular that way I
can add more items to the shop later and add more types of wellbeing
stats without modifying much code if any at all.

Now, I have to hook-up the native module that connects to the Health
Kit API I made earlier to the in-game currency.