Finished in-game currency addition
I completed the mechanic of viewing how many steps the player has
accumulated for the day and covering that to in-game currency.

Now, I just have to do some clean-up with styling, background images,
lowering the Pet Wellbeing stats at a reasonable speed, and saving the
Redux state to async local storage.